# public/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Orders Rest API"
  },
  "host": "",
  "basePath": "",
  "tags": [
    {
      "name": "orders",
      "description": "Orders Rest API"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/orders": {
      "post": {
        "tags": [
          "orders"
        ],
        "summary": "Create new order",
        "description": "Create new order",
        "operationId": "addOrder",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Order object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Orders"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid input request"
          }
        }
      },
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "Get orders list",
        "description": "Get orders list",
        "operationId": "listorder",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "description": "Current Page No.",
            "required": true,
            "default":1
          },
          {
            "in": "query",
            "name": "limit",
            "description": "No. of items to be returned",
            "required": true,
            "default":5
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid input request"
          }
        }
      }
    },
    "/orders/{orderId}": {
      "patch": {
        "tags": [
          "orders"
        ],
        "summary": "Update an order status",
        "description": "Take order",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "GUID",
            "description": "The Order Id Of specific order"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Order object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/OrderUpdate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid request parameters"
          },
          "409": {
            "description": "Already taken"
          },
          "404": {
            "description": "Order not found"
          }
        }
      }
    }
  },
  "definitions": {
    "Orders": {
      "type": "object",
      "required": [
        "origin",
        "destination"
      ],
      "properties": {
        "origin": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ObjectSchemaDefinition"
          }
        },
        "destination": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ObjectSchemaDefinition"
          }
        }
      }
    },
    "OrderUpdate": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "description": "Order status",
          "enum": [
            "TAKEN",
            "UNASSIGNED"
          ]
        }
      }
    },
    "ObjectSchemaDefinition": {
      "type": "string"
    }
  }
}